refactor(grow): extract late LLM phases to llm_phases.py
Move phases 8c (overlays), 9 (choices), 9b (fork beats), and 9c
(hub-spokes) from stage.py into _LLMPhaseMixin in llm_phases.py.
stage.py drops from 1,319 to 432 lines.
